eta_microservice/requirements.txt
Flask==3.0.0
psycopg2==2.9.9
numpy==1.26.1
pandas==2.1.2
geopy==2.4.0
python-dotenv==1.0.0